+++ src/natten/na2d.py
@@ -27,6 +27,7 @@ from torch import nn, Tensor
 from torch.nn.init import trunc_normal_
 
 from .context import is_fna_enabled
+from .experimental import na2d as experimental_na2d
 from .functional import na2d, na2d_av, na2d_qk
 from .types import CausalArg2DTypeOrDed, Dimension2DTypeOrDed
 from .utils import check_all_args, log
@@ -51,6 +52,7 @@ class NeighborhoodAttention2D(nn.Module):
         qk_scale: Optional[float] = None,
         attn_drop: float = 0.0,
         proj_drop: float = 0.0,
+        use_experimental_ops: bool = False,
     ):
         super().__init__()
         kernel_size_, dilation_, is_causal_ = check_all_args(
@@ -87,6 +89,8 @@ class NeighborhoodAttention2D(nn.Module):
         self.proj = nn.Linear(dim, dim)
         self.proj_drop = nn.Dropout(proj_drop)
 
+        self.use_experimental_ops = use_experimental_ops
+
     def forward(self, x: Tensor) -> Tensor:
         if x.dim() != 4:
             raise ValueError(
@@ -110,7 +114,7 @@ class NeighborhoodAttention2D(nn.Module):
                 .permute(3, 0, 1, 2, 4, 5)
             )
             q, k, v = qkv[0], qkv[1], qkv[2]
-            x = na2d(
+            x = (experimental_na2d if self.use_experimental_ops else na2d)(
                 q,
                 k,
                 v,
@@ -123,6 +127,10 @@ class NeighborhoodAttention2D(nn.Module):
             x = x.reshape(B, H, W, C)
 
         else:
+            if self.use_experimental_ops:
+                raise NotImplementedError(
+                    "Only fused NA is included in experimental support for torch.compile and torch's FLOP counter."
+                )
             qkv = (
                 self.qkv(x)
                 .reshape(B, H, W, 3, self.num_heads, self.head_dim)